Add durability config options for multiple items

+++ common/src/main/java/artifacts/item/ArtifactProperties.java
@@ -145,7 +145,7 @@ public final class ArtifactProperties {
         return component(DataComponents.BREAK_SOUND, breakSound);
     }
 
-    public ArtifactProperties damageOverTime(Value<Integer> damagePerSecond, EntityCondition condition) {
+    public ArtifactProperties damageOverTime(Value<Double> damagePerSecond, EntityCondition condition) {
         return component(ModDataComponents.DAMAGE_OVER_TIME.get(), new DamageOverTime(damagePerSecond, condition));
     }
